Remove opensource.apple.com from source_sites
This was done because the version tags we are using
in the PureDarwin github repositories do not follow
the same (unknown) versioning schema as the packages
on opensource.apple.com. It is entirely possible that
a conflict may occur.

Furthermore, the packages from opensource.apple.com
wouldn't work in the first place anyway, because the
github repos contain extensive fixes and patches.