Add fix_dtrace_collision patch
This is one of the silliest errors I have ever seen.
If a macro that contains a DTrace call is present on
the same line number in two different files, a symbol
collision will occur, resulting in a linker error. All
that is needed to fix it is add one or more blank/comment
lines to one of the files.