Fix parts of libsyscall build
This commit introduces a long list of headers
in the libc_headers subdirectory. These are required
by the libsyscall code, but are not used by it (since
that would be a layering violation). These headers would
usually be provided by the Xcode SDK, but we disable the
use of that.