Vendor iig outputs into source tree
Since recreating the functionality of iig will take time,
and we only use it to generate these output files, why
not include them in the source tree to remove the dependency?

iig has been replaced in the makefiles by the false command.
This should fail the build if it is ever unexpected used
anywhere else.