Remove -target flags from compiler command lines
This was for some reason causing the Clang engine in my editor
to use the syntax `-arch x86_64-apple-darwin11.0`. The value after
the -arch flag should be just 'x86_64', which it was set to elsewhere,
but that didn't work for some reason. I am changing this here for
the time being.