Remove shared/user differentiation in corecrypto
This was done in order to potentially share the corecrypto_shared
library between libcorecrypto, libcorecrypto_static, and the
corecrypto kext, but I am not certain if the algorithm code is
the exact same in both places. It took a while getting the kext
to build and boot; I'd rather not risk breaking it.