Set explicit file extension on dynamic libraries
On macOS, the file extension for dynamic libraries is
'.dylib'. On Linux, it is '.so'. This change should
aid in crosscompilation on Linux.

I would have submitted this change in a PR, but my branch
has become corrupted for some reason. GitHub does not
accept it, saying it is out of date with the main branch,
but there are no changes on the main branch I can rebase
over to solve the problem.