tools/xar: make xar buildable on Linux/glibc
This brings us a step closer to being able to build PureDarwin on Linux.

Some of the required symbols in glibc are only visible when specific
feature-test macros are defined; define them. Additionally, remove some
hardcoded feature macros from config.h and implement the checks in CMake.

Also remove a bunch of trailing whitespace because of editorconfig.